Refs #6781. Taking EntityId_t to its own header.
